2_SlidesTranslator/requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.0.380
langchain-core>=0.1.13
python-pptx>=0.6.21
python-dotenv>=1.0.0
asyncio